requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
altair==5.5.0
    # via streamlit
asgiref==3.9.1
    # via django
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.0
    # via streamlit
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via click
django==5.2.5
    # via -r src/requirements/requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
lxml==6.0.1
    # via python-docx
markupsafe==3.0.2
    # via jinja2
narwhals==2.2.0
    # via altair
numpy==2.3.2
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.2
    # via streamlit
pillow==11.3.0
    # via streamlit
protobuf==6.32.0
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pypdf2==3.0.1
    # via -r src/requirements/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r src/requirements/requirements.in
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r src/requirements/requirements.in
    #   streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sqlparse==0.5.3
    # via django
streamlit==1.49.0
    # via -r src/requirements/requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   python-docx
    #   referencing
    #   streamlit
tzdata==2025.2
    # via
    #   django
    #   pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
